Require Verified+1/2 by Zuul for all events
Zuulv3 was configured to enqueue changes to the gate pipeline
on patchset created and recheck comment events only when the
Verified+1 vote was present. This, however, didn't matched other
internal Zuul events like the "merge completed" event for example.
A global "require" stanza for the gate pipeline covers all of the
possible cases.
Other implication of this change is that the force-merge scenario
with opencontrail-admin account will no longer work.

Change-Id: I8a8ba8dcc6530aa5b4067b5d2b615169b23fa3d5