Support non-release branches in repo init
Change #43920 introduced checking out repo manifest from target branch
of the review. This breaks the repo init for projects that use
non-release branches (e.g. contrail/ocata in contrail-kolla-ansible).
This change reads the branch checked-out by zuul and uses it in repo
init. This way we leave the correct branch resolution to Zuul
(effectively RX.X for release branches and master for all the rest).

Change-Id: I2d680b4e1bb3b04ba3077e13cfdffa94d8685d18