Nightly - publish all images with non-sku tags
Originally, the only Docker image without an OpenStack-version specific tag
('master-123' instead of 'queens-master-123') was the contrail-go image.
After creation of the deployer containers we need to publish more images with
this tag type, so we're dropping the filter on contrail-go and instead match
all the existing images.

On request by bharatkk@juniper.net

Change-Id: I9102f234f20bc490fde2ecfffc2037df5e33ac03