Do not publish rocky
Since rocky should only be built and published
for `master` we can not have it as a dependent job
for publishing due to how we're currently storing
configuration (it would be missing from the job
graph due to Iebde4603ca8cc835d7007f00e0176e6e8e8548cf).

We can not have it configured for publishing due to
race conditions so for now it will have to be published
manually.

Related ticket: https://contrail-jws.atlassian.net/browse/JD-510

Change-Id: Id0919eb73485c8b3d8faebe3788787c7656f29ae